All features working!!
With this commit, al features are working!!! Including, kicks, bans,
temp-ban, console log, save and load console functions, click on map to
teleport target,etc.

The work is not concluded, has a lot of bugs and unoptimized code
(specially netcode). Functions are not compiled and does not support
admin levels.